requirements.txt
pillow
pytesseract
pyautogui
opencv-python-headless
requests
bs4
pandas
IPython
matplotlib 